api/requirements.txt
fastapi==0.110.0
openai>=1.24.0
pinecone-client>=3.0.0
python-dotenv==1.0.1
langchain==0.1.20
langchain-openai==0.1.7
langchain-community==0.0.38
pypdf==4.0.1
pydantic==2.6.1
python-multipart==0.0.9
tiktoken>=0.7.0
langchain-pinecone==0.1.0 